(sol): day 7, completed week 1

+++ day_7/day_7.cpp
@@ -2,9 +2,9 @@
 using namespace std;
 
 bool check_calibration(int64_t n, const vector<int64_t>& ns);
-int64_t cal(const vector<int64_t>& ns, const vector<size_t>& ops);
+int64_t cal(const vector<int64_t>& ns, const vector<int>& ops);
 bool check_calibration_with_concatenation(int64_t n, const vector<int64_t>& ns);
-int64_t cal_with_concatenation(const vector<int64_t>& ns, const vector<size_t>& ops);
+int64_t cal_with_concatenation(const vector<int64_t>& ns, const vector<int>& ops);
 int64_t n_len(int64_t n);
 
 int main() {
@@ -47,11 +47,11 @@ int main() {
 }
 
 bool check_calibration(int64_t n, const vector<int64_t>& ns) {
-    size_t l = ns.size() - 1;
-    vector<size_t> ops(l, 0);
-    size_t all = 1 << l; // 2^l
-    for (size_t i = 0; i < all; ++i) {
-        size_t j = i;
+    int l = ns.size() - 1;
+    vector<int> ops(l, 0);
+    int all = 1 << l; // 2^l
+    for (int i = 0; i < all; ++i) {
+        int j = i;
         for (auto& op : ops) {
             op = j % 2;
             j /= 2;
@@ -64,20 +64,20 @@ bool check_calibration(int64_t n, const vector<int64_t>& ns) {
     return false;
 }
 
-int64_t cal(const vector<int64_t>& ns, const vector<size_t>& ops) {
+int64_t cal(const vector<int64_t>& ns, const vector<int>& ops) {
     int64_t ans = ns[0];
-    for (size_t i = 0; i < ops.size(); ++i) {
+    for (int i = 0; i < ops.size(); ++i) {
         ans = (ops[i] == 0) ? ans + ns[i + 1] : ans * ns[i + 1];
     }
     return ans;
 }
 
 bool check_calibration_with_concatenation(int64_t n, const vector<int64_t>& ns) {
-    size_t l = ns.size() - 1;
-    vector<size_t> ops(l, 0);
-    size_t all = pow(3, l); // 3^l
-    for (size_t i = 0; i < all; ++i) {
-        size_t j = i;
+    int l = ns.size() - 1;
+    vector<int> ops(l, 0);
+    int all = pow(3, l); // 3^l
+    for (int i = 0; i < all; ++i) {
+        int j = i;
         for (auto& op : ops) {
             op = j % 3;
             j /= 3;
@@ -90,9 +90,9 @@ bool check_calibration_with_concatenation(int64_t n, const vector<int64_t>& ns) 
     return false;
 }
 
-int64_t cal_with_concatenation(const vector<int64_t>& ns, const vector<size_t>& ops) {
+int64_t cal_with_concatenation(const vector<int64_t>& ns, const vector<int>& ops) {
     int64_t ans = ns[0];
-    for (size_t i = 0; i < ops.size(); ++i) {
+    for (int i = 0; i < ops.size(); ++i) {
         if (ops[i] == 0) {
             ans += ns[i + 1];
         } else if (ops[i] == 1) {
